refactor: updated the folder structure and naming convention
