+++ src/components/App.js
@@ -1,9 +1,9 @@
 import React from 'react';
 
-import './App.css';
+import '../styles/scss/App.scss';
 
-import Display from './components/display/display.component';
-import ButtonPanel from './components/button-panel/button-panel.component';
+import Display from './Display';
+import ButtonPanel from './ButtonPanel';
 
 function App() {
   return (
