+++ src/components/Button.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import './button.styles.scss';
+import '../styles/scss/Button.scss';
 
 const isOperator = val => val === '+' || val === '-' || val === 'X' || val === '÷' || val === '=';
